% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_functions.R
\name{cat_header}
\alias{cat_header}
\title{Write Markdown headers}
\usage{
cat_header(text = "", level = 3)
}
\arguments{
\item{text}{text to put as header}

\item{level}{text level}
}
\value{
A Markdown-formatted heading
}
\description{
Writes the appropriate number of hash signs for the specified text level.
Intended for creating sections in an R Markdown document using for-loops.
}
