% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_barplot.R
\name{create_barplot}
\alias{create_barplot}
\title{Create bar charts for categorical variables}
\usage{
create_barplot(
  df,
  filter_var = NULL,
  filter_value = NULL,
  facet = FALSE,
  facet_var = NULL,
  plot_title,
  aes_x,
  aes_y = ..prop..,
  xlab,
  ylab = "Proportion"
)
}
\arguments{
\item{df}{An object of class tbl.df, data.frame or tbl.}

\item{filter_var}{Variable you want to filter the dataset on.}

\item{filter_value}{Value to filter for.}

\item{facet}{Boolean. Add a facet to the plot or not.}

\item{facet_var}{Variable to facet on.}

\item{plot_title}{A string. Plot title.}

\item{aes_x}{The variable to be displayed on the x-axis.}

\item{aes_y}{The variable to be displayed on the y-axis.}

\item{xlab}{x-axis label.}

\item{ylab}{y-axis label.}
}
\value{
A ggplot object. May add additional layers with a \code{+}.
}
\description{
Create bar charts for categorical variables
}
