% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kable_functions.R
\name{format_kable}
\alias{format_kable}
\title{Automatic kable formatting}
\usage{
format_kable(df, position = "center")
}
\arguments{
\item{df}{An object of class tbl.df, data.frame or tbl.}

\item{position}{Alignment of kable output as a string. Default value is \code{"center"}.}
}
\value{
A kable. Output may be piped into specific formatting preferences for each column.
}
\description{
Formats data frames and tibbles as kables.
}
