% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_functions.R
\name{formula_generator}
\alias{formula_generator}
\title{Formula generator}
\usage{
formula_generator(predictors, response)
}
\arguments{
\item{predictors}{A vector of predictor variables as strings}

\item{response}{A response variable as a string}
}
\value{
a formula of the form \code{y ~ x} as string
}
\description{
Writes a formula as a string.
}
\examples{
\dontrun{formula_generator(predictors, 'Cluster')}
}
