% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_category.R
\name{plot_category}
\alias{plot_category}
\title{Plotting categorical variables}
\usage{
plot_category(df, plot_type, var1, var2)
}
\arguments{
\item{df}{An object of class data.frame or tbl.}

\item{plot_type}{Type of plot to be created (boxplot or histogram).}

\item{var1}{Categorical variable in data frame to be used as x-axis in the
boxplot, or as the grouping variable in the histogram.}

\item{var2}{Categorical variable in data frame to be used as y-axis in the
boxplot, or as x-axis in the histogram in the histogram.}
}
\value{
A ggplot object. May add additional layers with a \code{+}.
}
\description{
A function that generates either a boxplot or histogram.
}
