% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_functions.R
\name{word_saver}
\alias{word_saver}
\title{Export data frame to Word document}
\usage{
word_saver(df, directory)
}
\arguments{
\item{df}{The data frame to be exported}

\item{directory}{The directory where the Word file should be saved as a string.}
}
\value{
Location of Word document created
}
\description{
Exports a data frame as a Word document in the directory specified
}
